fix: correct colors, styling and patternization
